Fix casting issue

+++ Roblox/Roblox.Website/Controllers/Internal/BypassController.cs
@@ -1913,15 +1913,15 @@ namespace Roblox.Website.Controllers
                     new 
                     {
                         Name = "Test",
-                        Type = Models.Assets.Type.Hat,
+                        Type = (int)Models.Assets.Type.Hat,
                         TargetId = 164,
                         Asset = new
                         {
                             Id = 164,
-                            TypeId = Models.Assets.Type.Hat,
+                            TypeId = (int)Models.Assets.Type.Hat,
                             Name = "Name",
                             Description = "Name",
-                            CreatorType = CreatorType.User,
+                            CreatorType = (int)CreatorType.User,
                             CreatorTargetId = 1,
                             Created = "2017-03-31T12:16:46.547",
                             Updated = "2017-03-31T12:16:46.547",
